main interactive mode
