+++ tkcore/term/CMakeLists.txt
@@ -1,8 +1,6 @@
 add_executable(tkey
         utils/term_utils.cpp
         utils/interactive.cpp
-        cmd_processing.cpp
-        interactive.cpp
         storage_v1.cpp
         storage_v2.cpp
         main.cpp
