--- tkcore/term/interactive.cpp
@@ -1,122 +1,0 @@
-//
-// Created by panda on 21.01.24.
-//
-
-#include "cmd_processing.h"
-#include "public/storage2/storage.h"
-#include "public/key_storage_v1.h"
-#include "def_header.h"
-#include "utils/term_utils.h"
-
-using namespace thekey_term;
-using namespace term;
-using namespace std;
-
-#define COLUMN_WIDTH 40
-static char ident = '\t';
-
-static void printInteractiveModeHelp();
-
-static void findStorage();
-
-static void createStorage();
-
-void interactive() {
-    cout << "TheKey interactive mode" << endl;
-
-    int exitFlag = 0;
-    while (!exitFlag) {
-        auto cmd = ask_from_term("> ");
-
-        if (cmd == "q" || cmd == "exit" || cmd == "quit")
-            return;
-
-        if (cmd == "h" || cmd == "help") {
-            printInteractiveModeHelp();
-            continue;
-        }
-
-        if (cmd == "find") {
-            findStorage();
-            continue;
-        }
-
-        if (cmd == "create") {
-            createStorage();
-            continue;
-        }
-
-        if (cmd == "login") {
-            auto path = ask_from_term("input path : ");
-            login(path);
-            continue;
-        }
-
-        cerr << "cmd not support " << cmd << endl;
-    }
-
-}
-
-static void printInteractiveModeHelp() {
-    cout << "TheKey - (storageVersion. " << TERM_VERSION << ") cryp/encrypt your secure passwords storages" << endl;
-    cout << endl;
-    cout << ident << "Commands:" << endl;
-
-    cout << ident;
-    cout.width(COLUMN_WIDTH);
-    cout << std::left << "help or h";
-    cout << "help for commands" << endl;
-
-    cout << ident;
-    cout.width(COLUMN_WIDTH);
-    cout << std::left << "find";
-    cout << "find storage" << endl;
-
-    cout << ident;
-    cout.width(COLUMN_WIDTH);
-    cout << std::left << "create";
-    cout << "create storage" << endl;
-
-    cout << ident;
-    cout.width(COLUMN_WIDTH);
-    cout << std::left << "login";
-    cout << "login storage" << endl;
-}
-
-void findStorage() {
-    auto path = ask_from_term("input path: ");
-    thekey_term::findStorages(path);
-}
-
-void createStorage() {
-    auto path = ask_from_term("input path: ");
-    auto version = ask_int_from_term("input storage version: ");
-    if (version > 2) {
-        cerr << "version not support " << version << " last version support " << 2 << endl;
-        return;
-    }
-    auto name = ask_from_term("input name: ");
-    auto dest = ask_from_term("input dest: ");
-
-    int error = 0;
-    switch (version) {
-        case 1: {
-            error = thekey_v1::createStorage(
-                    {.file = path, .storageVersion = uint(version), .name = name, .description = dest});
-            break;
-        }
-        case 2: {
-            error = thekey_v2::createStorage(
-                    {.file = path, .storageVersion = uint(version), .name = name, .description = dest});
-            break;
-        }
-    }
-
-    if (error) {
-        cout << "error to create storage " << error << endl;
-    } else {
-        cout << "storage created successfully" << endl;
-    }
-}
-
-
