+++ tkcore/term/main.cpp
@@ -1,57 +1,108 @@
 //
 // Created by panda on 06.06.2020.
 //
-#include "cmd_processing.h"
 #include "def_header.h"
+#include "utils/Interactive.h"
+#include "utils/term_utils.h"
+#include "storage_v1.h"
+#include "storage_v2.h"
+#include "public/storage2/storage.h"
 
-using namespace thekey_term;
+#ifdef __ANDROID__
+namespace fs = std::__fs::filesystem;
+#else
+namespace fs = std::filesystem;
+#endif
 
-extern void interactive();
-
+using namespace term;
+using namespace std;
+using namespace thekey;
 
 int main(int argc, char **argv) {
-    const char *short_options = ":hf:l:";
-    const struct option long_options[] = {
-            {"help",  no_argument,       NULL, 'h'},
-            {"find",  optional_argument, NULL, 'f'},
-            {"login", required_argument, NULL, 'l'},
-            {NULL, 0,                    NULL, 0}
-    };
+    auto it = Interactive();
+    it.welcomeText = string("TheKey - ")
+                     + "cryp/encrypt your secure passwords storages\n"
+                     + "Designed by Andrey Kuzubov / klee0kai";
+    it.byeText = "bye";
+    it.helpTitle = string("TheKey Main interactive mode ")
+                   + "(app version " + TERM_VERSION + ") ";
 
+    it.cmd({"f", "find"}, "find available storages in folder", []() {
+        auto folder = ask_from_term("input path: ");
+        cout << "Available crypted storages in folder: " << fs::absolute(folder) << endl;
+        thekey::findStorages(folder, [](const Storage &storage) {
+            cout << "-------------------------------------" << endl;
+            cout << "storagePath: " << storage.file << endl;
+            cout << "name: " << storage.name << endl;
+            cout << "version: " << storage.storageVersion << endl;
+            cout << "desc: " << storage.description << endl;
+            cout << endl;
+        });
+        cout << "-------------------------------------" << endl;
+    });
 
-    if (argc <= 1) {
-        interactive();
-        return 0;
-    }
-
-    char ch = 0;
-    while (ch = getopt_long(argc, argv, short_options, long_options, NULL), ch != -1) {
-        switch (ch) {
-            case 'h':
-                thekey_term::printHelp();
-                return 0;
-            case 'f':
-                thekey_term::findStorages(optarg);
-                return 0;
-            case 'l':
-                thekey_term::login(optarg);
-                return 0;
-            case ':':
-                switch (optopt) {
-                    case 'f':
-                        thekey_term::findStorages(optarg);
-                        return 0;
-                    default:
-                        std::cerr << "invalid option: " << (char) optopt << std::endl;
-                        return EXIT_FAILURE;
-
-                }
-            default:
-                std::cerr << "invalid option: " << (char) optopt << std::endl;
-                return EXIT_FAILURE;
+    it.cmd({"c", "create"}, "create new storage", []() {
+        auto path = ask_from_term("input path: ");
+        auto version = ask_int_from_term("input storage version: ");
+        if (version > 2) {
+            cerr << "version not support " << version << " last version support " << 2 << endl;
+            return;
         }
-    }
+        auto name = ask_from_term("input name: ");
+        auto dest = ask_from_term("input dest: ");
 
+        int error = 0;
+        switch (version) {
+            case 1: {
+                error = thekey_v1::createStorage(
+                        {
+                                .file = path,
+                                .storageVersion = uint(version),
+                                .name = name,
+                                .description = dest
+                        });
+                break;
+            }
+            case 2: {
+                error = thekey_v2::createStorage(
+                        {
+                                .file = path,
+                                .storageVersion = uint(version),
+                                .name = name,
+                                .description = dest
+                        });
+                break;
+            }
+        }
+
+        if (error) {
+            cout << "error to create storage " << error << endl;
+        } else {
+            cout << "storage created successfully" << endl;
+        }
+    });
+
+    it.cmd({"l", "login"}, "enter encrypted storage", []() {
+        auto filePath = ask_from_term("input path : ");
+        auto storageInfo = thekey::storage(filePath);
+        if (!storageInfo) {
+            cerr << "can't open file " << filePath << endl;
+            return;
+        }
+        switch (storageInfo->storageVersion) {
+            case 1:
+                thekey_term_v1::login(filePath);
+                return;
+            case 2:
+                thekey_term_v2::login(filePath);
+                return;
+            default:
+                cerr << "storage version " << storageInfo->storageVersion << " not supported " << filePath << endl;
+                return;
+        }
+    });
+
+    it.loop();
 
     return 0;
 }
